Add files via upload
In this project I used Recurrent Neural Network to predict the Google's Stock. Libraries that I used in the project are, numpy for mathematical computations and pre processing the data,  matplotlib was used to visualize the data and the main RNN was built using keras framework.  The model was trained for 100 epochs with batch size of 32.